rename entities field to better names + use long instead of for date

+++ app/src/main/java/com/ssmmhh/jibam/persistence/entities/TransactionEntity.kt
@@ -25,10 +25,8 @@ data class TransactionEntity(
     //category id exactly id
     @ColumnInfo(name = COLUMN_CATEGORY_ID)
     val cat_id: Int,
-    //int can handle the time till 1/19/2038, 6:44:07 AM in millisecond
     @ColumnInfo(name = "date")
-    //TODO ("use long for data")
-    val date: Int,
+    val date: Long,
 ) {
     companion object {
         const val COLUMN_CATEGORY_ID = "categoryId"
