+++ app/src/main/java/com/ssmmhh/jibam/ui/main/transaction/addedittransaction/detailedittransaction/state/DetailEditTransactionViewState.kt
@@ -88,7 +88,7 @@ class SubmitButtonState() {
 
     }
 
-    fun onDateChange(newDate: Int) {
+    fun onDateChange(newDate: Long) {
         _doesDateChange.value = defaultTransaction?.date != newDate
 
     }
