+++ app/src/main/java/com/ssmmhh/jibam/ui/main/transaction/addedittransaction/inserttransaction/InsertTransactionFragment.kt
@@ -358,7 +358,7 @@ constructor(
             money = money,
             memo = binding.edtMemo.text.toString(),
             cat_id = category.id,
-            date = (calender.timeInMillis).div(1_000).toInt()
+            date = (calender.timeInMillis).div(1_000)
         )
     }
